refactor(main): 파일 구조 변경 및 PYTHONPATH 환경변수 추가
