+++ README.md
@@ -4,6 +4,12 @@
 - codespace 로 웹 서버 띄우기 가능
 - ui 는 html 로 구현
 
+### 변경사항 생겼을시
+```
+python manage.py migrate
+python runserver
+```
+
 ## Pieces
 - [ ] 자기가 움직일 수 있는 범위인가
 
@@ -27,10 +33,11 @@
 - [x] 움직일 기물과 움직일 위치의 좌표를 사용자로 부터 입력 받음
 
 ## Test
-- [ ] 테스트 코드 실행시 src/main 디렉토리에 접근하지 못함
-- [ ] 임시로 아래 코드를 모든 테스트 파일에 작성하여 해결함
+- [X] 테스트 코드 실행시 src/main 디렉토리에 접근하지 못함
+- [X] 임시로 아래 코드를 모든 테스트 파일에 작성하여 해결함
 ``` 
 import sys
-sys.path.append("./src/main/chess/chess")
+sys.path.append("./src/main/chess")
 ```
-- [ ] 추후 리펙토링 필요
+- [X] 추후 리펙토링 필요
+- `PYTHONPATH` 환경변수를 추가하여 해결
