+++ src/main/chess/pieces/Bishop.py
