+++ src/main/chess/pieces/NonePiece.py
