+++ src/main/chess/pieces/Queen.py
