+++ src/main/server/chess_api/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class MainConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "chess"
+    name = "chess_api"
