+++ src/main/server/config/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "chess",
+    "chess_api",
     "rest_framework",
 ]
