+++ src/main/server/config/urls.py
@@ -19,5 +19,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path('', include('chess.urls')),
+    path('', include('chess_api.urls')),
 ]
