+++ src/main/server/resource/static/js/chessboard.js
