+++ src/test/chess/pieces/KingTest.py
@@ -1,7 +1,5 @@
 import unittest
-import sys
-sys.path.append("./src/main/chess/chess")
-from pieces.King import King
+from chess.pieces.King import King
 
 class KingTest(unittest.TestCase):
