+++ src/test/chess/pieces/PieceTest.py
@@ -1,7 +1,5 @@
 import unittest
-import sys
-sys.path.append("./src/main/chess")
-from pieces.Pawn import Pawn
+from chess.pieces.Pawn import Pawn
 
 class PieceTest(unittest.TestCase):
